+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 
 ## Unreleased
 
+* Fixes an issue where `<template>` elements were being ignored by generate and build. 
+* Fixes an issue where `rosey check` would error when run without a source
+
 ## v2.0.4 (March 16, 2023)
 
 * `<title>`, `<script>` and `<style>` elements will no longer be wordwrapped when the `--wrap` flag is used.